refactor(outbox): S30 Phase F — share envelope + delete legacy postgres relay
Collapse the three duplicate copies of the outbox wire envelope
(adapters/postgres outboxMessage, runtime/eventbus outboxWireMessage,
adapters/rabbitmq outboxWireMessage) onto runtime/outbox.WireMessage and
route all inbound decoding through outbox.UnmarshalEnvelope. RabbitMQ keeps
its PascalCase-Entry legacy fallback local since that path is transport-
specific backwards compatibility, not part of the shared envelope contract.

Delete the legacy adapter-layer outbox relay now that cmd/core-bundle and
the fullchain integration test are wired to runtime/outbox.NewRelay. The
SQL text, constants, and helper functions that outbox_store.go relied on
(relayDB, status constants, truncateError / sanitizeError) are split out
to adapters/postgres/outbox_db.go, and the unit-test mock types are moved
to adapters/postgres/outbox_mock_test.go to keep them alive independently
of the deleted outbox_relay_test.go.

Removes roughly 2.8kloc of adapter code (outbox_relay*.go,
safe_relay_collector*.go) against ~300 lines of envelope-sharing edits.

Resolves S28 (envelope consolidation) in the same PR as S30.
