+++ adapters/rabbitmq/subscriber.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/ghbvf/gocell/kernel/outbox"
 	"github.com/ghbvf/gocell/pkg/errcode"
+	outboxrt "github.com/ghbvf/gocell/runtime/outbox"
 )
 
 // errSubscriptionLost is a sentinel error returned by subscribeOnce when the
@@ -603,67 +604,31 @@ func (s *Subscriber) Close() error {
 // Wire format deserialization
 // ---------------------------------------------------------------------------
 
-// outboxWireMessage is the wire envelope produced by the three-phase relay.
-// Fields use camelCase JSON tags.
-//
-// NOTE: adapters/postgres/outbox_relay.go defines an identical outboxMessage
-// for serialization — keep the two structs in sync when modifying fields.
-type outboxWireMessage struct {
-	ID            string            `json:"id"`
-	AggregateID   string            `json:"aggregateId,omitempty"`
-	AggregateType string            `json:"aggregateType,omitempty"`
-	EventType     string            `json:"eventType"`
-	Topic         string            `json:"topic,omitempty"`
-	Payload       json.RawMessage   `json:"payload"`
-	Metadata      map[string]string `json:"metadata,omitempty"`
-	CreatedAt     time.Time         `json:"createdAt"`
-}
-
 // unmarshalDelivery deserializes a broker message body into an outbox.Entry.
-// It first tries the new outboxWireMessage envelope, then falls back to the
-// legacy full outbox.Entry format for backward compatibility.
+// It delegates envelope detection to outboxrt.UnmarshalEnvelope (canonical
+// WireMessage format). When that returns an empty entry (non-envelope payload),
+// it falls back to the legacy PascalCase full outbox.Entry JSON format used by
+// older queues, preserving backward compatibility with pre-envelope messages.
 //
-// Discriminator: In the new wire format, payload is embedded JSON (starts
-// with '{' or '[' as json.RawMessage). In legacy format, outbox.Entry.Payload
-// is []byte which json.Marshal encodes as base64 (starts with '"'). Go's
-// json.Unmarshal does case-insensitive key matching, so we cannot rely on
-// PascalCase vs camelCase to distinguish formats — we must check the payload
-// shape instead.
+// Discriminator: outboxrt.UnmarshalEnvelope requires non-empty ID + EventType
+// AND an embedded JSON payload (starts with '{' or '['). If any condition fails
+// the function returns a fresh stamped entry — we detect this by checking whether
+// the returned entry's ID has the "evt-" prefix (fallback path). In that case we
+// attempt the legacy PascalCase unmarshal before accepting the fallback.
+//
+// Note: the "evt-" prefix check is a deliberate heuristic; real relay-produced
+// IDs are plain UUIDs without that prefix.
 func unmarshalDelivery(body []byte) (outbox.Entry, error) {
-	var msg outboxWireMessage
-	if err := json.Unmarshal(body, &msg); err == nil && msg.ID != "" && msg.EventType != "" && isEmbeddedJSON(msg.Payload) {
-		return outbox.Entry{
-			ID:            msg.ID,
-			AggregateID:   msg.AggregateID,
-			AggregateType: msg.AggregateType,
-			EventType:     msg.EventType,
-			Topic:         msg.Topic,
-			Payload:       []byte(msg.Payload),
-			Metadata:      msg.Metadata,
-			CreatedAt:     msg.CreatedAt,
-		}, nil
-	}
-
-	// Fallback: legacy full Entry (PascalCase, Payload is base64-encoded []byte).
-	var entry outbox.Entry
-	if err := json.Unmarshal(body, &entry); err != nil {
-		return outbox.Entry{}, fmt.Errorf("unmarshal delivery: %w", err)
+	entry, _ := outboxrt.UnmarshalEnvelope("", body)
+	// UnmarshalEnvelope returns a freshly-stamped entry (ID starts with "evt-")
+	// when the body does not match the WireMessage discriminator. In that case try
+	// the legacy PascalCase full Entry JSON (backward-compatible RabbitMQ path).
+	if len(entry.ID) >= 4 && entry.ID[:4] == "evt-" {
+		var legacy outbox.Entry
+		if err := json.Unmarshal(body, &legacy); err != nil {
+			return outbox.Entry{}, fmt.Errorf("unmarshal delivery: %w", err)
+		}
+		return legacy, nil
 	}
 	return entry, nil
-}
-
-// isEmbeddedJSON returns true if the raw JSON value is an object or array
-// (new wire format), as opposed to a base64 string (legacy format).
-func isEmbeddedJSON(raw json.RawMessage) bool {
-	for _, b := range raw {
-		switch b {
-		case ' ', '\t', '\n', '\r':
-			continue
-		case '{', '[':
-			return true
-		default:
-			return false
-		}
-	}
-	return false
 }
